reject resync start on misconfigured replication rules
we expect resync to start on buckets with replication
rule ExistingObjects enabled, if not we reject such
calls.
